Stop maintaining the lang files as separate files.
Given what you can see on the dev streams or blog posts, the gui file
was probably already significantly changed in the upcoming v1.2.

Maintaining this file (and the other langfiles) separatedly will
be a nightmare when v1.2 comes out since we will have to update it
manually.

They are now complete and maintained in the big pack file, so that
jsontr can detect the stale translations, making things both simpler
and easier.