Add a note indicating this mod is pretty much not a good starter example.
Apparently people have trouble using Localize-Me, and it's almost as if
they think I can bite people over the internet if they try to ask for help.
I can assure you that I can't and don't want to.

In either case, French is bloated. it patches fonts, it patches numbers,
it implements non-breaking space characters, and the author lives in
a copyright maximalist country. So no, it's not a good template to use.