Fix variable references to landmarks.
The checker was supposedly supposed to check them, and checked them it
was. Except the checker would only accepts area.(.*).landmarks.name.(.*)
which is actually invalid. The game want area.(.*).landmark.name.(.*)
which only differs by a single 's' character, but the checker would
reject them as invalid, of course.

To add to the confusion, that reference points to
database.areas[\1].landmarks[\2].name, with 'landmarks' ending in 's'.

So pull in the fixed checker and fix those texts.