Stop encrapting files.
At 2020-02-15 10:00 (UTC+??), in the official CrossCode Discord server
in the #modding channel, R.D. and Lachsen explicitely granted modders
the right to distribute the "game script" along with translations.

Since that's simpler for everyone and people keeps complaining that
Localize-Me-Tools is too complicated to use, decrapt everything.  If
only it could attract more pull requests...

And since we are rewriting packs entirely, sort them by game order,
instead of sorting lexicographically, which breaks the ordering of
conversations in arrays (the classical "1" < "10" < "2" problem...).

As for the legalese, while it is unclear whether i have a partial
copyright on the derived work, it seems that i don't have to pay taxes
to SOFIA if I distribute freely, unlike for sound/music, where almost
everything is taxed.