Update tools and add some stale texts.
At this point, well... Let's just stop pretending.  The great apocalypse
is coming and it will spare no-one.  This commit is merely nothing but a
futile attempt at containing it, by acting on dubious speculation to
prepare for the, for our, inevitable doom.

It adds even more stale texts, as if it was the solution to everything.
Sure, a lot of "stale" texts were added, but they were nothing, nothing
but mods that were taken as granted, but not from the original game.

No, this time, the plot is even more sinister.  These texts, they are
the real deal.  Actual game text.  Which now have a stale translation.

These texts... They are lurking in dark corners, so much that you aren't
even supposed to see them... yet.  And by the time they come out, these
forgotten stale texts, perhaps... they might grant us a chance.

A chance to make it out alive, and to finish this once and for all.