Perform manual migration to 1.4.0-3
That's right, to stand on the shoulder of giants, just read the
fucking mnaual !

More seriously, fix all the texts that the automatic migration messed
up.  This probably does not mean that everything from the non-DLC still
works, because it's still possible that some stuff have been moved and
haven't been recognised.