Translate more things from the base game.
Also, update to 1.4.0-4.  This was horribly hard, given that a grand
total of three texts were changed, pfew !  Never again.

Also, update the README and config example while we are at it.

The "DLC" is still not translated.  And by "DLC", i mean only the
additional files.  And not the regular maps where DLC-exclusive quests
were added.