Patch some ADD_MSG_PERSON that don't use lang labels.
This happens only in two obscure places.
One is after a post-DLC quest in the heat village (the one with the
delightful pre-DLC description) which is fine because we don't localize
the name of the person anyway.
The other one is a hidden-by-default quest in the cold village which is
more visible because we actually localize the name of the person.  Also
the translation is bad because, well, the cold village was among the
first texts to be translated.