Fix ccmod.json whitespace.
ccmod.json was initially implemented by ccloader3.  It used to have no
specification (as usual) so the only point of reference was other
ccloader3 mods like crosscode-ru and the ccloader3 source code.

ccloader2 eventually supported ccmod.json.  Lately, a new ccmod.json
specification is being written¹ which adds two new useful fields: "tags"
and "authors".

For now, it is not incompatible with ccloader3, which is nice because
this mod is both a ccloader2 and ccloader3 mod.

[1]: https://github.com/CCDirectLink/CCModDB/blob/39e0db315c72869c88ebb2767fcc2cefd1123ebc/CCMOD-STANDARD.md