Github actions: update ccloader2 and localize-me version
The previous ccloader2 now crashes on linux, and there are a few fixes
for ccloader3 in localize-me.  Also switch back to 0.x for localize-me,
because the state of modloaders haven't changed.

So many years have passed, and github sure did a good job at
enshitification. Let's hope these workflows works the first time i use
them, because otherwise it's going to be a steamy pile of shit commits.